# MainApp/docs/swagger.yaml
definitions:
  echo.HTTPError:
    properties:
      message: {}
    type: object
...
  models.User:
    properties:
      avatar:
        type: integer
      email:
        type: string
      first_name:
        type: string
      id:
        readOnly: true
        type: integer
      last_name:
        type: string
      nick_name:
        type: string
      password:
        type: string
    required:
    - email
    - first_name
    - last_name
    - nick_name
    - password
    type: object
...
  pkg.Response:
    properties:
      body: {}
    type: object
host: 89.208.197.127:8080
info:
  contact: {}
  title: WS Swagger API
  version: "1.0"
paths:
...
  /friends/{user_id}:
    get:
      description: get user's friends
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success get profile
          schema:
            allOf:
            - $ref: '#/definitions/pkg.Response'
            - properties:
                body:
                  items:
                    $ref: '#/definitions/models.User'
                  type: array
              type: object
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "401":
          description: no cookie
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "404":
          description: user doesn't exist
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "405":
          description: Method Not Allowed
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: SelectFriends
      tags:
      - friends
  /friends/add/{friend_id}:
    post:
      description: add friend
      parameters:
      - description: Friend ID
        in: path
        name: friend_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: friend added
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "401":
          description: no cookie
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "403":
          description: invalid csrf
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "404":
          description: friend or user doesn't exist
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "405":
          description: Method Not Allowed
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "409":
          description: friend already exists
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: AddFriend
      tags:
      - friends
  /friends/check/{friend_id}:
    get:
      description: check friend
      parameters:
      - description: Friend ID
        in: path
        name: friend_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success get profile
          schema:
            allOf:
            - $ref: '#/definitions/pkg.Response'
            - properties:
                body:
                  type: boolean
              type: object
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "401":
          description: no cookie
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "404":
          description: item doesn't exist
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "405":
          description: Method Not Allowed
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: CheckIsFriend
      tags:
      - friends
  /friends/delete/{friend_id}:
    delete:
      description: delete friend
      parameters:
      - description: Friend ID
        in: path
        name: friend_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: friend deleted, body is empty
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "401":
          description: no cookie
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "403":
          description: invalid csrf
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "404":
          description: friend/user/friendship doesn't exist
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "405":
          description: Method Not Allowed
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: DeleteFriend
      tags:
      - friends
...
swagger: "2.0"